requirements/tests.txt
-e .
psycopg2
pytest-django
tox
